[FIX] quality_control: do not delete QC for unfinished pickings
When splitting a picking (thus creating a backorder), the quality
checks that are still in state 'none' are deleted because Odoo
assumes that the old picking is done and the QCs are not needed
anymore.

This fix ensures that QCs for split pickings that are still in
progress remain.

opw-5193424

closes odoo/enterprise#101532

X-original-commit: 0844194292437cc0bc243603f8622f836998c28e

+++ quality_control/models/stock_picking.py
@@ -75,7 +75,8 @@ class StockPicking(models.Model):
         for backorder in res:
             # Do not link the QC of move lines with quantity of 0 in backorder.
             backorder.move_line_ids.filtered(lambda ml: not float_is_zero(ml.quantity, precision_rounding=ml.product_uom_id.rounding)).check_ids.picking_id = backorder
-            backorder.backorder_id.check_ids.filtered(lambda qc: qc.quality_state == 'none').sudo().unlink()
+            if backorder.backorder_id.state in ('done', 'cancel'):
+                backorder.backorder_id.check_ids.filtered(lambda qc: qc.quality_state == 'none').sudo().unlink()
             backorder.move_ids._create_quality_checks()
         return res
